crud for cart and purchase is working

+++ src/controllers/cart.controller.js
@@ -1,44 +1,58 @@
 const catchError = require('../utils/catchError');
 const Cart = require('../models/Cart');
+const Product = require('../models/Product');
+const Category = require('../models/Category');
 
 const getAll = catchError(async(req, res) => {
-    const results = await Cart.findAll();
-    return res.json(results);
-});
-
-const create = catchError(async(req, res) => {
-    const result = await Cart.create(req.body);
-    return res.status(201).json(result);
-});
-
-const getOne = catchError(async(req, res) => {
-    const { id } = req.params;
-    const result = await Cart.findByPk(id);
-    if(!result) return res.sendStatus(404);
+    const user = req.user.id;
+    const results = await Cart.findAll({
+        where: { userId },
+        include: [{
+            //include product
+            model: Product,
+            attributes: { exclude:[ "updatedAt" , "createdAt" ] },
+            include: {
+                include: Category, 
+                attributes: ['name']
+            }
+        }]
+    });
     return res.json(result);
 });
 
+const create = catchError(async(req, res) => {
+    const userId = req.user.id
+    const { quantity, productId  } = req.body
+    const newBody = { userId, quantity, productId }
+    const result = await Cart.create(newBody);
+    return res.status(201).json(result);
+});
+
+
 const remove = catchError(async(req, res) => {
     const { id } = req.params;
-    const result = await Cart.destroy({ where: {id} });
+    const userId = req.user.id
+    const result = await Cart.destroy({ where: {id, userId} });
     if(!result) return res.sendStatus(404);
     return res.sendStatus(204);
 });
 
 const update = catchError(async(req, res) => {
+    const userId = req.user.id
     const { id } = req.params;
+    const { quantity } = req.body
     const result = await Cart.update(
-        req.body,
-        { where: {id}, returning: true }
+        { quantity },
+        { where: { id, userId }, returning: true }
     );
     if(result[0] === 0) return res.sendStatus(404);
     return res.json(result[1][0]);
 });
 
 module.exports = {
+    
     getAll,
     create,
-    getOne,
     remove,
     update
 }
